feat(cloudfront): resolved comments
