+++ prowler/providers/aws/services/cloudfront/cloudfront_service.py
@@ -44,6 +44,12 @@ class CloudFront(AWSService):
                                         )
                                         .get("OriginSslProtocols", {})
                                         .get("Items", []),
+                                        origin_access_control=origin.get(
+                                            "OriginAccessControlId", ""
+                                        ),
+                                        s3_origin_config=origin.get(
+                                            "S3OriginConfig", {}
+                                        ),
                                     )
                                 )
                             distribution = Distribution(
@@ -159,6 +165,8 @@ class Origin(BaseModel):
     domain_name: str
     origin_protocol_policy: str
     origin_ssl_protocols: list[str]
+    origin_access_control: Optional[str]
+    s3_origin_config: Optional[dict]
 
 
 class Distribution(BaseModel):
@@ -172,5 +180,4 @@ class Distribution(BaseModel):
     geo_restriction_type: Optional[GeoRestrictionType]
     origins: list[Origin]
     web_acl_id: str = ""
-    origin_access_control: Optional[bool]
     tags: Optional[list] = []
